Moved up the logic of the 'SSL untrusted' dialog to FileActivity to be shown in all refreshes
